+++ lib/Assertion/Ast/ExpressionNode.php
@@ -2,6 +2,8 @@
 
 namespace PhpBench\Assertion\Ast;
 
+use PhpBench\Expression\Ast\Node;
+
 interface ExpressionNode extends Node
 {
 }
