--- lib/Assertion/Ast/Node.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace PhpBench\Assertion\Ast;
-
-interface Node
-{
-}
