--- lib/Assertion/Ast/NumberNode.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace PhpBench\Assertion\Ast;
-
-interface NumberNode extends ExpressionNode
-{
-    public function value();
-}
