+++ lib/Expression/Ast/BinaryOperatorNode.php
@@ -2,7 +2,7 @@
 
 namespace PhpBench\Expression\Ast;
 
-use PhpBench\Assertion\Ast\Node;
+use PhpBench\Expression\Ast\Node;
 
 class BinaryOperatorNode implements Node
 {
