+++ lib/Expression/Ast/Node.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace PhpBench\Expression\Ast;
+
+interface Node
+{
+}
