+++ lib/Expression/Ast/NumberNode.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace PhpBench\Expression\Ast;
+
+interface NumberNode extends Node
+{
+    /**
+     * @return float|integer
+     */
+    public function value();
+}
