+++ lib/Expression/Parselet/BinaryOperatorParselet.php
@@ -2,10 +2,9 @@
 
 namespace PhpBench\Expression\Parselet;
 
-use PhpBench\Assertion\ArithmeticNode;
-use PhpBench\Assertion\Ast\Node;
-use PhpBench\Assertion\Token;
-use PhpBench\Assertion\Tokens;
+use PhpBench\Expression\Ast\Node;
+use PhpBench\Expression\Token;
+use PhpBench\Expression\Tokens;
 use PhpBench\Expression\Ast\BinaryOperatorNode;
 use PhpBench\Expression\InfixParselet;
 use PhpBench\Expression\Parser;
