+++ tests/Unit/Assertion/Printer/NodePrinterTest.php
@@ -5,11 +5,11 @@ namespace PhpBench\Tests\Unit\Assertion\Printer;
 use Generator;
 use PhpBench\Assertion\Ast\Comparison;
 use PhpBench\Assertion\Ast\DisplayAsNode;
-use PhpBench\Assertion\Ast\FloatNode;
-use PhpBench\Assertion\Ast\FunctionNode;
-use PhpBench\Assertion\Ast\IntegerNode;
+use PhpBench\Expression\Ast\FloatNode;
+use PhpBench\Expression\Ast\FunctionNode;
+use PhpBench\Expression\Ast\IntegerNode;
 use PhpBench\Assertion\Ast\MemoryValue;
-use PhpBench\Assertion\Ast\Node;
+use PhpBench\Expression\Ast\Node;
 use PhpBench\Assertion\Ast\ParenthesizedExpressionNode;
 use PhpBench\Assertion\Ast\PercentageValue;
 use PhpBench\Assertion\Ast\PropertyAccess;
