+++ tests/Unit/Expression/Parselet/BinaryOperatorParseletTest.php
@@ -5,7 +5,7 @@ namespace PhpBench\Tests\Unit\Expression\Parselet;
 use Generator;
 use PHPUnit\Framework\TestCase;
 use PhpBench\Assertion\ArithmeticNode;
-use PhpBench\Assertion\Ast\IntegerNode;
+use PhpBench\Expression\Ast\IntegerNode;
 use PhpBench\Expression\Ast\BinaryOperatorNode;
 use PhpBench\Tests\Unit\Expression\ParseletTestCase;
